fix: invoke mapNodesToParents lazily
