+++ lib/style.js
@@ -1,7 +1,8 @@
 import * as csstree from 'css-tree';
 import * as csswhat from 'css-what';
 import { syntax } from 'csso';
-import { matches, visit } from './xast.js';
+import { matches } from './xast.js';
+import { visit } from './util/visit.js';
 import {
   attrsGroups,
   inheritableAttrs,
