+++ lib/style.test.js
@@ -1,5 +1,5 @@
 import { collectStylesheet, computeStyle } from './style.js';
-import { visit } from './xast.js';
+import { visit } from './util/visit.js';
 import { parseSvg } from './parser.js';
 
 /**
