+++ lib/svgo.js
@@ -1,7 +1,8 @@
 import { builtinPlugins } from './builtin.js';
 import { encodeSVGDatauri } from './svgo/tools.js';
 import { invokePlugins } from './svgo/plugins.js';
-import { mapNodesToParents, querySelector, querySelectorAll } from './xast.js';
+import { querySelector, querySelectorAll } from './xast.js';
+import { mapNodesToParents } from './util/map-nodes-to-parents.js';
 import { parseSvg } from './parser.js';
 import { stringifySvg } from './stringifier.js';
 import { VERSION } from './version.js';
