+++ lib/svgo/css-select-adapter.js
@@ -1,3 +1,5 @@
+import { mapNodesToParents } from '../util/map-nodes-to-parents.js';
+
 /** @type {Required<import('css-select').Options<import('../types.js').XastNode & { children?: any }, import('../types.js').XastElement>>['adapter']['isTag']} */
 const isTag = (node) => {
   return node.type === 'element';
@@ -69,12 +71,17 @@ const findOne = (test, elems) => {
 };
 
 /**
- * @param {Map<import('../types.js').XastNode, import('../types.js').XastParent>} parents
+ * @param {import('../types.js').XastParent} relativeNode
+ * @param {Map<import('../types.js').XastNode, import('../types.js').XastParent>=} parents
  * @returns {Required<import('css-select').Options<import('../types.js').XastNode & { children?: any }, import('../types.js').XastElement>>['adapter']}
  */
-export function createAdapter(parents) {
+export function createAdapter(relativeNode, parents) {
   /** @type {Required<import('css-select').Options<import('../types.js').XastNode & { children?: any }, import('../types.js').XastElement>>['adapter']['getParent']} */
   const getParent = (node) => {
+    if (!parents) {
+      parents = mapNodesToParents(relativeNode);
+    }
+
     return parents.get(node) || null;
   };
