+++ lib/svgo/plugins.js
@@ -1,4 +1,4 @@
-import { visit } from '../xast.js';
+import { visit } from '../util/visit.js';
 
 /**
  * Plugins engine.
