+++ lib/xast.js
@@ -2,42 +2,15 @@ import { is, selectAll, selectOne } from 'css-select';
 import { createAdapter } from './svgo/css-select-adapter.js';
 
 /**
- * @param {Map<import('./types.js').XastNode, import('./types.js').XastParent>} parents
+ * @param {import('./types.js').XastParent} relativeNode
+ * @param {Map<import('./types.js').XastNode, import('./types.js').XastParent>=} parents
  * @returns {import('css-select').Options<import('./types.js').XastNode & { children?: any }, import('./types.js').XastElement>}
  */
-function createCssSelectOptions(parents) {
+function createCssSelectOptions(relativeNode, parents) {
   return {
     xmlMode: true,
-    adapter: createAdapter(parents),
+    adapter: createAdapter(relativeNode, parents),
   };
-}
-
-/**
- * Maps all nodes to their parent node recursively.
- *
- * @param {import('./types.js').XastParent} node
- * @returns {Map<import('./types.js').XastNode, import('./types.js').XastParent>}
- */
-export function mapNodesToParents(node) {
-  /** @type {Map<import('./types.js').XastNode, import('./types.js').XastParent>} */
-  const parents = new Map();
-
-  for (const child of node.children) {
-    parents.set(child, node);
-    visit(
-      child,
-      {
-        element: {
-          enter: (child, parent) => {
-            parents.set(child, parent);
-          },
-        },
-      },
-      node,
-    );
-  }
-
-  return parents;
 }
 
 /**
@@ -46,12 +19,8 @@ export function mapNodesToParents(node) {
  * @param {Map<import('./types.js').XastNode, import('./types.js').XastParent>=} parents
  * @returns {import('./types.js').XastChild[]} All matching elements.
  */
-export const querySelectorAll = (
-  node,
-  selector,
-  parents = mapNodesToParents(node),
-) => {
-  return selectAll(selector, node, createCssSelectOptions(parents));
+export const querySelectorAll = (node, selector, parents) => {
+  return selectAll(selector, node, createCssSelectOptions(node, parents));
 };
 
 /**
@@ -60,12 +29,8 @@ export const querySelectorAll = (
  * @param {Map<import('./types.js').XastNode, import('./types.js').XastParent>=} parents
  * @returns {?import('./types.js').XastChild} First match, or null if there was no match.
  */
-export const querySelector = (
-  node,
-  selector,
-  parents = mapNodesToParents(node),
-) => {
-  return selectOne(selector, node, createCssSelectOptions(parents));
+export const querySelector = (node, selector, parents) => {
+  return selectOne(selector, node, createCssSelectOptions(node, parents));
 };
 
 /**
@@ -74,45 +39,8 @@ export const querySelector = (
  * @param {Map<import('./types.js').XastNode, import('./types.js').XastParent>=} parents
  * @returns {boolean}
  */
-export const matches = (node, selector, parents = mapNodesToParents(node)) => {
-  return is(node, selector, createCssSelectOptions(parents));
-};
-
-export const visitSkip = Symbol();
-
-/**
- * @param {import('./types.js').XastNode} node
- * @param {import('./types.js').Visitor} visitor
- * @param {any=} parentNode
- */
-export const visit = (node, visitor, parentNode) => {
-  const callbacks = visitor[node.type];
-  if (callbacks?.enter) {
-    // @ts-expect-error hard to infer
-    const symbol = callbacks.enter(node, parentNode);
-    if (symbol === visitSkip) {
-      return;
-    }
-  }
-  // visit root children
-  if (node.type === 'root') {
-    // copy children array to not lose cursor when children is spliced
-    for (const child of node.children) {
-      visit(child, visitor, node);
-    }
-  }
-  // visit element children if still attached to parent
-  if (node.type === 'element') {
-    if (parentNode.children.includes(node)) {
-      for (const child of node.children) {
-        visit(child, visitor, node);
-      }
-    }
-  }
-  if (callbacks?.exit) {
-    // @ts-expect-error hard to infer
-    callbacks.exit(node, parentNode);
-  }
+export const matches = (node, selector, parents) => {
+  return is(node, selector, createCssSelectOptions(node, parents));
 };
 
 /**
