+++ lib/xast.test.js
@@ -1,4 +1,5 @@
-import { detachNodeFromParent, visit, visitSkip } from './xast.js';
+import { detachNodeFromParent } from './xast.js';
+import { visit, visitSkip } from './util/visit.js';
 
 /**
  * @param {import('./types.js').XastElement[]} children
