+++ plugins/removeUselessStrokeAndFill.js
@@ -1,4 +1,5 @@
-import { detachNodeFromParent, visit, visitSkip } from '../lib/xast.js';
+import { detachNodeFromParent } from '../lib/xast.js';
+import { visit, visitSkip } from '../lib/util/visit.js';
 import { collectStylesheet, computeStyle } from '../lib/style.js';
 import { hasScripts } from '../lib/svgo/tools.js';
 import { elemsGroups } from './_collections.js';
